feat(session-memory): wire observatory into extraction + compaction paths
MemoryExtractionService now records every skip/terminal outcome onto
the observatory ring (opt-in via `.with_observatory(arc)`). Records
distinguish trigger (init_gate / growth_gate / error_override), carry
the selector model actually used, preserve content preview (clipped,
Unicode-safe) and narrative section names for Persisted outcomes, and
surface latency.

Compaction side: `MemoriaCompactParams` grows two optional fields —
`turn_number` + `observatory`. Every facts-first injection produces one
InjectionRecord tagged with pressure, InjectionLevel, injected_chars,
FactsSummary (turn/plan/files/errors/last-error-preview), staleness
signals, retrieved Memoria memory refs (id + type + score; no
content), and the narrative sections kept after cross-validation.

Cache safety (explicit constraint): the injected string and LLM
message payload are bit-identical to pre-change — session_memory_
protocol.rs is untouched, the new helper runs after injection is
built and only writes to observatory. No prompt hashes shift.

Context safety: previews clipped to 200 chars, retrieved memories
carry only metadata, rings bounded at 32/16. Observatory reads never
re-enter system prompt or rolling cache — they surface only through
the introspect tool result path (added in a follow-up commit).

Wiring points: `ServerAgenticLoopHost::compact_messages_via_memoria`
reads `state.memory_extraction_service.observatory()` and passes it
via `MemoriaContext` into the compaction call. Bridge and offline
paths leave observatory=None — they're not used in production server
runs where the ring matters.

TDD / unhappy first:
  * 5 service-side tests: skip-below-init-gate records as Skipped,
    had_error path records as ErrorOverride even when skipped,
    persisted-rule-based produces Persisted record with preview,
    None observatory stays silent, empty session_id suppresses record.
  * 4 compaction-side tests: records full injection + facts summary
    + retrieved memory metadata, empty memories under retrieve
    failure, L0Only pressure records injected_chars=0, None
    observatory is silent.

Small refactors pulled in:
  * `ExtractionArtifacts::{Persisted,LlmFailedPersistedFallback}`
    now carry the persisted `content` so the service can extract
    section names + preview without re-parsing. PersistFailed arm
    stays content-less.
  * `MemoryExtractionService::memoria_breaker_state()` +
    `pending_drain()` accessors — needed for the introspect subtopic
    next commit.
  * 27 `MemoriaCompactParams` construct sites updated with the two
    new fields defaulting to `turn_number: 0, observatory: None`.

+++ rust/crates/runtime/src/session_memory/observatory.rs
@@ -186,7 +186,7 @@ pub struct StalenessSignals {
 pub struct RetrievedMemoryRef {
     pub memory_id: String,
     pub memory_type: String,
-    pub score: Option<f32>,
+    pub score: Option<f64>,
 }
 
 /// One compaction-time injection event, post-hoc.
@@ -447,9 +447,7 @@ mod tests {
         // And the whole thing must still be valid UTF-8 (implicit via
         // String, explicit via the round-trip).
         assert!(
-            clipped
-                .chars()
-                .all(|c| c == '🦀' || c == '…'),
+            clipped.chars().all(|c| c == '🦀' || c == '…'),
             "no partial codepoints"
         );
     }
